wrapp events routes with auth

+++ server/routes/events.js
@@ -1,5 +1,6 @@
 const express = require('express')
 const router = express.Router()
+const { getTokenDecoder } = require('authenticare/server')
 
 const { saveNewEvent,
   getAllEvents,
@@ -7,32 +8,43 @@ const { saveNewEvent,
   addUserToEvent,
   getUsersForEvent } = require('../Db/projectDb')
 
-router.get('/', (req, res) => {
-  getAllEvents(req.body)
-    .then(events => res.status(200).json(events))
-    .catch(() => res.status(500).send('DATABASE ERROR'))
+router.get('/', getTokenDecoder(), (req, res) => {
+  if (req.user) {
+    getAllEvents(req.body)
+      .then(events => res.status(200).json(events))
+      .catch(() => res.status(500).send('DATABASE ERROR'))
+  } else {
+    res.status(500).send('authentication token not provided')
+  }
 })
 
-router.post('/:id', (req, res) => {
-  console.log('route')
-  const id = Number(req.params.id)
-  saveNewEvent(id, req.body)
-    .then((ids) => {
-      return res.status(200).json(ids[0])
-    })
-    .catch(err => console.log(err))
+router.post('/:id', getTokenDecoder(), (req, res) => {
+  if (req.user) {
+    const id = Number(req.params.id)
+    saveNewEvent(id, req.body)
+      .then((ids) => {
+        return res.status(200).json(ids[0])
+      })
+      .catch(err => console.log(err))
+  } else {
+    res.status(500).send('authentication token not provided')
+  }
 })
 
-router.delete('/', (req, res) => {
-  const id = Number(req.body)
-  deleteEvent(id)
-    .then(() => {
-      res.sendStatus(200)
-      return null
-    })
-    .catch(error => {
-      res.status(500).send('DATABASE ERROR' + error.message)
-    })
+router.delete('/', getTokenDecoder(), (req, res) => {
+  if (req.user) {
+    const id = Number(req.body)
+    deleteEvent(id)
+      .then(() => {
+        res.sendStatus(200)
+        return null
+      })
+      .catch(error => {
+        res.status(500).send('DATABASE ERROR' + error.message)
+      })
+  } else {
+    res.status(500).send('authentication token not provided')
+  }
 })
 
 module.exports = router
